Migrate `getClosestParent` to TypeScript
