+++ packages/js-toolkit/helpers/index.ts
@@ -1,7 +1,7 @@
-export { default as createApp } from './createApp.js';
 export * from './getDirectChildren.js';
+export { default as createApp } from './createApp.js';
+export { default as getClosestParent } from './getClosestParent.js';
 export { default as getInstanceFromElement } from './getInstanceFromElement.js';
 export { default as importOnInteraction } from './importOnInteraction.js';
 export { default as importWhenIdle } from './importWhenIdle.js';
 export { default as importWhenVisible } from './importWhenVisible.js';
-export { default as getClosestParent } from './getClosestParent.js';
